refactor: 코드리뷰 피드백 반영
- config 파일에 transform 관련 변수 추가
- DetailPost에 title 잘못 들어간 것 수정
- LocationnCarousel의 return문 내에 변수로 수정
- RelatedPost의 if-else를 삼항 연산자로 수정
- RelatedPostComment의 위치 변경 (RelatedPost의 하위 폴더로 이동)
- 닫는 태그를 축약형으로 통일
